Enhanced Mandelbrot visualisation and fixed complex number calculation
Implemented a new way to visualise Mandelbrot set by dividing a panel into squares and painting them based on the intensity calculated from Mandelbrot data in PnlMandelbrot.java. Both Mandelbrot.java and Complex.java classes were refactored. In Mandelbrot class, removed the static 'step' variable and added system out print for debugging purposes in the 'eval' method. Corrected the 'sqr' and 'moduleSquared' methods in Complex class to properly calculate the square and the square of the modulus of a complex number. This improves accuracy and aids in future computations.