Increase viewport size and add mouse interaction in Mandelbrot App
The size of the Mandelbrot application view is adjusted to 600x500 from 300x200 in MandelbrotApp.java. This is to give users a larger window for considering the fractal. Added MouseListener and respective methods in PnlMandelbrot.java, allowing users to interact with the fractal, such as zooming in and out. Removed a debug print line in Mandelbrot.java to tidy up the code. Performance and interaction improvement should provide a better user experience.